+++ R/piat.R
@@ -23,7 +23,7 @@ piat <- function(num_items = 25L,
                  take_training = TRUE,
                  label = "PIAT",
                  feedback = piat.feedback.no_score(dict = dict),
-                 media_dir = "http://media.gold-msi.org/test_materials/PIAT/1-0-1/mp4",
+                 media_dir = "https://media.gold-msi.org/test_materials/PIAT/1-0-1/mp4",
                  dict = piat::piat_dict) {
   stopifnot(is.scalar.character(label), is.scalar.numeric(num_items),
             is.scalar.logical(take_training), is.scalar.character(media_dir))